requirements.txt
flask==2.0.1
scikit-learn==0.24.2
numpy==1.21.0
pandas==1.3.0
pymongo==4.3.3
python-dotenv==0.19.0